add pre-prepared data for SANS
