+++ 4-reduction/sans_utils.py
@@ -6,6 +6,8 @@ import scipp as sc
 
 from load import load_ascii, load_nexus
 
+from utils import fetch_data  # noqa: F401
+
 
 def load_sans(
     path: str,
